+++ source/extensions/omni.isaac.ros2_bridge/docs/CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGELOG
 Changed
 ---------
 - Output data types to vectord and quatd in ROS2 Odometry, Raw TF publisher nodes and Twist subscriber node
+- Added dropdown menu and validation for encoding input in ROS2 image publisher node
 
 [0.8.1] - 2022-05-02
 ========================